refactor init_model and unit test
