+++ mmpose/apis/legacy/__init__.py
@@ -1,0 +1,20 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .inference import (collect_multi_frames, inference_bottom_up_pose_model,
+                        inference_top_down_pose_model, init_pose_model,
+                        process_mmdet_results, vis_pose_result)
+from .inference_3d import (extract_pose_sequence, inference_interhand_3d_model,
+                           inference_mesh_model, inference_pose_lifter_model,
+                           vis_3d_mesh_result, vis_3d_pose_result)
+from .inference_tracking import get_track_id, vis_pose_tracking_result
+from .test import multi_gpu_test, single_gpu_test
+from .train import init_random_seed, train_model
+
+__all__ = [
+    'train_model', 'init_pose_model', 'inference_top_down_pose_model',
+    'inference_bottom_up_pose_model', 'multi_gpu_test', 'single_gpu_test',
+    'vis_pose_result', 'get_track_id', 'vis_pose_tracking_result',
+    'inference_pose_lifter_model', 'vis_3d_pose_result',
+    'inference_interhand_3d_model', 'extract_pose_sequence',
+    'inference_mesh_model', 'vis_3d_mesh_result', 'process_mmdet_results',
+    'init_random_seed', 'collect_multi_frames'
+]
