+++ mmpose/models/heads/heatmap_heads/heatmap_head.py
@@ -247,7 +247,7 @@ class HeatmapHead(BaseHead):
         """Predict results from features."""
 
         batch_heatmaps = self.forward(feats)
-        preds = self.decode(batch_heatmaps, batch_data_samples, test_cfg)
+        preds = self.decode(batch_heatmaps, batch_data_samples)
 
         # Whether to visualize the predicted heatmaps
         if test_cfg.get('output_heatmaps', False):
